Refactor About screen into a separate package.
