+++ zeapp/android/src/main/java/de/berlindroid/zeapp/ZeMainActivity.kt
@@ -1,6 +1,5 @@
 package de.berlindroid.zeapp
 
-import android.content.Context
 import android.content.Intent
 import android.graphics.Bitmap
 import android.net.Uri
@@ -14,14 +13,12 @@ import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.imePadding
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.lazy.LazyColumn
@@ -68,7 +65,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.StrokeCap
@@ -108,12 +104,12 @@ import de.berlindroid.zeapp.zeui.WeatherEditorDialog
 import de.berlindroid.zeapp.zeui.ZeCameraEditor
 import de.berlindroid.zeapp.zeui.ZeImageDrawEditorDialog
 import de.berlindroid.zeapp.zeui.ZeNavigationPad
+import de.berlindroid.zeapp.zeui.zeabout.ZeAbout
 import de.berlindroid.zeapp.zeui.zetheme.ZeBadgeAppTheme
 import de.berlindroid.zeapp.zeui.zetheme.ZeBlack
 import de.berlindroid.zeapp.zeui.zetheme.ZeWhite
 import de.berlindroid.zeapp.zevm.ZeBadgeViewModel
 import de.berlindroid.zeapp.zevm.copy
-import de.berlindroid.zekompanion.getPlatform
 import kotlinx.coroutines.launch
 import timber.log.Timber
 import android.content.res.Configuration as AndroidConfig
@@ -274,7 +270,7 @@ private fun ZeScreen(vm: ZeBadgeViewModel, modifier: Modifier = Modifier) {
                     },
                     content = { paddingValues ->
                         if (isShowingAbout) {
-                            ZeAbout(paddingValues, vm, LocalContext.current)
+                            ZeAbout(paddingValues)
                         } else {
                             ZePages(
                                 paddingValues = paddingValues,
@@ -406,60 +402,6 @@ private fun ZeDrawerContent(
                     painter = painterResource(id = R.drawable.ic_update),
                     onClick = onGotoReleaseClick,
                 )
-            }
-        }
-    }
-}
-
-@Composable
-private fun ZeAbout(
-    paddingValues: PaddingValues,
-    vm: ZeBadgeViewModel,
-    context: Context,
-) {
-    val lines by vm.lines.collectAsState()
-
-    ZeSurface(
-        modifier = ZeModifier
-            .fillMaxSize()
-            .padding(paddingValues)
-            .padding(ZeDimen.Half),
-    ) {
-        Column {
-            ZeText(
-                text = "${lines.count()} contributors",
-                modifier = Modifier.padding(8.dp),
-                style = MaterialTheme.typography.bodyMedium,
-                fontSize = 24.sp,
-            )
-            ZeText(
-                text = "Running on '${getPlatform()}'.",
-            )
-            ZeLazyColumn {
-                items(lines) { line ->
-                    ZeRow(
-                        verticalAlignment = Alignment.CenterVertically,
-                    ) {
-                        val email = line.substring(line.indexOf('<').plus(1), line.lastIndexOf('>')).trim()
-                        ZeText(
-                            text = line.substring(0, line.indexOf('<')).trim(),
-                            color = MaterialTheme.colorScheme.onSurface,
-                            modifier = Modifier.padding(8.dp),
-                            style = MaterialTheme.typography.bodyMedium,
-                            fontSize = 18.sp,
-                        )
-                        ZeIcon(
-                            painter = painterResource(id = R.drawable.email),
-                            contentDescription = "Send random page to badge",
-                            Modifier
-                                .size(20.dp, 20.dp)
-                                .clickable {
-                                    val intent = Intent(Intent.ACTION_VIEW, Uri.parse("mailto:$email"))
-                                    context.startActivity(intent)
-                                },
-                        )
-                    }
-                }
             }
         }
     }
