+++ zeapp/android/src/main/java/de/berlindroid/zeapp/zevm/ZeBadgeViewModel.kt
@@ -35,7 +35,6 @@ class ZeBadgeViewModel @Inject constructor(
     private val preferencesService: ZePreferencesService,
     private val clipboardService: ZeClipboardService,
     private val getTemplateConfigurations: GetTemplateConfigurations,
-    contributorsService: ZeContributorsService,
 ) : ViewModel() {
 
     private val _uiState: MutableStateFlow<ZeBadgeUiState> = MutableStateFlow(getInitialUIState())
@@ -494,8 +493,7 @@ class ZeBadgeViewModel @Inject constructor(
         }
     }
 
-    val lines: StateFlow<List<String>> = contributorsService.contributors()
-        .stateIn(viewModelScope, SharingStarted.Lazily, initialValue = emptyList())
+
 
     private fun getInitialUIState(): ZeBadgeUiState =
         ZeBadgeUiState(
